requirements.txt
python-telegram-bot==20.7
playwright==1.40.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
pillow==10.1.0
python-magic==0.4.27
chardet==5.2.0
aiofiles==23.2.1
async-timeout==4.0.3
urllib3==2.0.7
requests==2.31.0
python-dotenv==1.0.0
